% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathSimR-package.R
\docType{package}
\name{PathSimR-package}
\alias{PathSimR}
\alias{PathSimR-package}
\title{PathSimR: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
