% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor_cores_required.R
\name{processor_cores_required}
\alias{processor_cores_required}
\title{Calculate the number of cores to use, which is limited by both the number
available, and the number of replications required}
\usage{
processor_cores_required(replications, cores = parallel::detectCores())
}
\arguments{
\item{replications}{integer. Number of simulation replications required}

\item{cores}{integer. Number of processor cores available}
}
\value{
integer. The number of cores to use (1 minimum)
}
\description{
Calculate the number of cores to use, which is limited by both the number
available, and the number of replications required
}
